+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Logger using standard logging module
 - "--log-level" command line argument (more info in "--help")
-- Changelog
+- "log_level" constructor parameter, defaults to "critical"
+- Created changelog
 
 ### Changed
 - All print statements now handled by logger